% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_diff_expr.R
\name{rna_diff_expr}
\alias{rna_diff_expr}
\title{RNAseq Differential Expressison}
\usage{
rna_diff_expr(count_table, design_table, method = "DESeq2")
}
\description{
This function allows you to identify
differentially expressed genes.
}
\examples{

}
