% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_diff_expr.R
\name{rna_diff_expr}
\alias{rna_diff_expr}
\title{RNAseq Differential Expressison}
\usage{
rna_diff_expr(count_file, design_file, method = "DEseq2")
}
\description{
This function allows you to identify
differentially expressed genes.
}
\examples{
rna_diff_expr(count_file, design_file, method="DEseq")
rna_diff_expr(count_file, design_file, method="DEseq2")
rna_diff_expr(count_file, design_file, method="edgeR")
rna_diff_expr(count_file, design_file, method="limma-voom")
rna_diff_expr(count_file, design_file, method="sleuth")
rna_diff_expr(count_file, design_file, method="bayseq")
rna_diff_expr(count_file, design_file, method="NOIseq")
rna_diff_expr(count_file, design_file, method="EBseq")
rna_diff_expr(count_file, design_file, method="SAMseq")
}
