% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_diff_expr.R
\name{rna_diff_expr}
\alias{rna_diff_expr}
\title{RNAseq Differential Expressison}
\usage{
rna_diff_expr(count_table, design_table, method = "DESeq2")
}
\description{
This function allows you to identify
differentially expressed genes.
}
\examples{
rna_diff_expr(count_table, design_table, method="DESeq")
rna_diff_expr(count_table, design_table, method="DESeq2")
rna_diff_expr(count_table, design_table, method="edgeR")
rna_diff_expr(count_table, design_table, method="limma-voom")
rna_diff_expr(count_table, design_table, method="sleuth")
rna_diff_expr(count_table, design_table, method="bayseq")
rna_diff_expr(count_table, design_table, method="NOIseq")
rna_diff_expr(count_table, design_table, method="EBseq")
rna_diff_expr(count_table, design_table, method="SAMseq")
}
