% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_visualize.R
\name{rna_visualize}
\alias{rna_visualize}
\title{RNAseq Visualize}
\usage{
rna_visualize(data_table, method = "hist", lib = "base", col,
  design_table, gene)
}
\description{
Visualize
https://www.rdocumentation.org/packages/edgeR/versions/3.14.0/topics/plotSmear
}
\examples{
rna_visualize(data_table, method="hist")
rna_visualize(data_table, method="counts")
rna_visualize(data_table, method="MA")
rna_visualize(data_table, method="smear")
rna_visualize(data_table, method="MDS")
rna_visualize(data_table, method="PCA")
rna_visualize(data_table, method="BCV")
rna_visualize(data_table, method="dispersions")
rna_visualize(data_table, method="sparsity")
rna_visualize(data_table, method="volcano")

rna_visualize(data_table, method="hist", lib="base")
rna_visualize(data_table, method="hist", lib="ggplot")
}
