% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_visualize.R
\name{rna_visualize}
\alias{rna_visualize}
\title{RNAseq Visualize}
\usage{
rna_visualize(data, method = "hist", lib = "base")
}
\description{
Visualize
}
\examples{
rna_visualize(data,method="MDSplot")
https://www.rdocumentation.org/packages/edgeR/versions/3.14.0/topics/plotSmear
rna_visualize(data,method="", lib="base")
rna_visualize(data,method="", lib="ggplot")

rna_visualize(data,method="hist")
rna_visualize(data,method="boxplot")
rna_visualize(data,method="density")
rna_visualize(data,method="smear")
rna_visualize(data,method="cluster-expt")
rna_visualize(data,method="cluster-gene")
rna_visualize(data,method="MA")
rna_visualize(data,method="MDS")
rna_visualize(data,method="PCA")
rna_visualize(data,method="BCV")
rna_visualize(data,method="volcano")
rna_visualize(data,method="venn")
}
